+++ CHANGELOG.md
@@ -9,5 +9,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - PostgreSqlInstall
   - Install Postgres with specified features
-- PostgreSqlScript
+- Added PostgreSqlScript resource
   - Run T-SQL Scripts against Postgres